updated file structure, removed some comments
