+++ client/graphs/radialTree.jsx
@@ -1,10 +1,9 @@
 import React, { Component, createRef } from 'react';
 import * as d3 from 'd3';
-import './graphv4.scss';
 import treeData from './treeData';
-import { project } from './helperFunctions';
+import { project, diagonal } from './helperFunctions';
 
-class TestGraph extends Component {
+class RadialTree extends Component {
   constructor(props) {
     super(props);
     this.state = {
@@ -25,20 +24,6 @@ class TestGraph extends Component {
   componentDidMount() {
     const { diameter, duration, margin, width, height } = this.state;
     let i = 0;
-
-    // defining the angle of the link?
-    function diagonal(d) {
-      return (
-        'M' +
-        project(d.x, d.y) +
-        'C' +
-        project(d.x, (d.y + d.parent.y) / 2) +
-        ' ' +
-        project(d.parent.x, (d.y + d.parent.y) / 2) +
-        ' ' +
-        project(d.parent.x, d.parent.y)
-      );
-    }
 
     // grabbing from DOM
     const svg = d3.select(this.myRef.current);
@@ -110,8 +95,7 @@ class TestGraph extends Component {
         .attr('x', 10) // ! (d) => (d.children || d._children ? -13 : 13)) was putting the text on top of the <g> so it threw off the clicking
         .attr('text-anchor', 'start')
         .text((d) => d.data.name)
-        .style('fill-opacity', 1e-6)
-        .style('color', 'white');
+        .style('fill-opacity', 1e-6);
 
       // we are merging the original spot to the child point (overrwriting the objects)
       const childPoint = startingPoint.merge(node);
@@ -216,4 +200,4 @@ class TestGraph extends Component {
   }
 }
 
-export default TestGraph;
+export default RadialTree;
