+++ client/pages/homePage.tsx
@@ -1,12 +1,10 @@
 import React from 'react'
-import TestGraph from '../d3/testgraph.jsx';
 
 
 export default function homePage() {
   return (
     <div>
       <h1>HOME PAGE</h1>
-      <TestGraph />
     </div>
   )
 };
