+++ client/styles/index.scss
@@ -1,3 +1,5 @@
+@import './radialTree.scss';
+
 body {
   background-color: #282a36;
   color: white;
