+++ client/styles/radialTree.scss
@@ -9,7 +9,7 @@
 
 .node text {
   font: 12px sans-serif;
-  color: white !important;
+  //color: white !important;
 }
 
 .link {
